[Codegen] Add pattern to hoist expand_shape out of forall
