+++ compiler/src/iree/compiler/Codegen/Common/test/BUILD.bazel
@@ -128,6 +128,7 @@ iree_lit_test_suite(
             "tile_and_distribute_workgroups_using_forall.mlir",
             "tile_large_tensors.mlir",
             "transform_buffer_opt.mlir",
+            "transform_collapse_forall_dest.mlir",
             "transform_copy_operand.mlir",
             "transform_flatten_forall.mlir",
             "transform_hoist_forall.mlir",
